+++ src/Shiny.Notifications/Notification.cs
@@ -13,7 +13,7 @@ namespace Shiny.Notifications
         /// <summary>
         /// This will be different per platform
         /// </summary>
-        public static string? CustomSoundFilePath { get; set; }
+        public static NotificationSound DefaultSound { get; set; } = NotificationSound.DefaultSystem;
 
         /// <summary>
         /// You do not have to set this - it will be automatically set from the library if you do not supply one
@@ -61,5 +61,10 @@ namespace Shiny.Notifications
         /// Options specific to windows (Uwp)
         /// </summary>
         public UwpOptions? Windows { get; set; } = new UwpOptions();
+
+        /// <summary>
+        /// The sound you wish to use on your notification - defaults to Notification.DefaultSound
+        /// </summary>
+        public NotificationSound Sound { get; set; } = DefaultSound;
     }
 }