+++ ChangeLog.md
@@ -1,3 +1,14 @@
 ﻿# Change Log
 
-## Not maintained while beta is running
+## 1.0.1
+[Enhancement][Breaking] New way to set notification sounds which allows you to use system sounds - Notification.CustomSoundFilePath has been removed
+
+## 1.0.0
+Initial Release for
+* Shiny.Core
+* Shiny.Locations
+* Shiny.Notifications
+* Shiny.Net.Http
+* Shiny.Sensors
+* Shiny.SpeechRecognition
+* Shiny.Testing